+++ src/bin/games/genius_square/models/arrays/mod.rs
@@ -1,3 +1,5 @@
 /// Models for handling arrays
 
 pub mod models;
+
+pub use models::BinArray;
