+++ src/bin/games/genius_square/models/board/mod.rs
@@ -1,3 +1,5 @@
 /// Models for handling game board
 
 pub mod models;
+
+pub use models::GameBoard;
