+++ src/bin/games/genius_square/models/constants/mod.rs
@@ -4,3 +4,9 @@ pub mod board;
 pub mod dice;
 pub mod enums;
 pub mod pieces;
+
+pub use board::*;
+pub use dice::*;
+pub use enums::*;
+#[allow(unused)]
+pub use pieces::*;
