+++ src/bin/games/genius_square/models/dice/mod.rs
@@ -2,3 +2,6 @@
 
 pub mod methods;
 pub mod models;
+
+pub use models::Die;
+pub use methods::roll_dice;
