--- src/bin/games/genius_square/models/pieces/methods.rs
@@ -1,14 +1,0 @@
-/// ----------------------------------------------------------------
-/// IMPORTS
-/// ----------------------------------------------------------------
-
-// use super::constants;
-
-/// ----------------------------------------------------------------
-/// METHODS
-/// ----------------------------------------------------------------
-
-#[allow(unused)]
-pub fn method() {
-    panic!("not implemented");
-}
