+++ src/bin/games/genius_square/models/pieces/mod.rs
@@ -1,4 +1,5 @@
 /// Models for handling pieces
 
-pub mod methods;
 pub mod models;
+
+pub use models::Piece;
